+++ packages/primeng/src/index.ts
@@ -1,4 +1,5 @@
 export * from './lib/ngx-form-engine-primeng.component';
+export * from './lib/provide-ngx-json-forms-primeng';
 export * from './lib/ngx-json-form-stepper.component';
 export * from './lib/ngx-json-form-debug.component';
 export * from './lib/signature-pad.component';